android: move e-mail helpers to lib
Given it can be used at the other app too
