+++ android/App/src/test/kotlin/com/darakeon/dfm/extensions/NavigationTest.kt
@@ -1,20 +1,15 @@
 package com.darakeon.dfm.extensions
 
 import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
-import com.darakeon.dfm.R
 import com.darakeon.dfm.base.BaseActivity
 import com.darakeon.dfm.testutils.BaseTest
-import com.darakeon.dfm.testutils.TestException
 import com.darakeon.dfm.testutils.context.getCalledName
 import com.darakeon.dfm.testutils.execute
 import com.darakeon.dfm.utils.activity.mockContext
 import com.darakeon.dfm.welcome.WelcomeActivity
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.CoreMatchers.containsString
-import org.hamcrest.CoreMatchers.hasItem
-import org.hamcrest.CoreMatchers.startsWith
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Assert.assertTrue
 import org.junit.Before
@@ -104,37 +99,6 @@ class NavigationTest: BaseTest() {
 	}
 
 	@Test
-	fun contact() {
-		`when`(activity.getString(R.string.dfm_mail_address))
-			.thenReturn("ADDRESS")
-
-		`when`(activity.getString(R.string.contact_subject))
-			.thenReturn("SUBJECT")
-
-		`when`(activity.getString(R.string.contact_body))
-			.thenReturn("BODY")
-
-		activity.contact()
-
-		assertThat(calledIntent?.action, `is`(Intent.ACTION_SENDTO))
-		assertThat(calledIntent?.data, `is`(Uri.parse("mailto:")))
-
-		val extras = calledIntent?.extras ?: Bundle()
-
-		assertThat(
-			extras.getStringArray(Intent.EXTRA_EMAIL)?.toList(),
-			hasItem("ADDRESS")
-		)
-
-		assertThat(extras.getString(Intent.EXTRA_SUBJECT), `is`("SUBJECT"))
-
-		assertThat(
-			extras.getString(Intent.EXTRA_TEXT),
-			startsWith("BODY")
-		)
-	}
-
-	@Test
 	fun goToSettings() {
 		val intent = mock(Intent::class.java)
 		`when`(activity.intent).thenReturn(intent)
@@ -172,57 +136,5 @@ class NavigationTest: BaseTest() {
 		val extras = calledIntent?.extras ?: Bundle()
 		assertThat(extras["__parent"]?.toString(), containsString("BaseActivity"))
 		assertThat(extras["test"]?.toString(), `is`("passed"))
-	}
-
-	@Test
-	fun composeErrorEmail() {
-		`when`(activity.getString(R.string.error_mail_title))
-			.thenReturn("SUBJECT")
-
-		`when`(activity.getString(R.string.dfm_mail_address))
-			.thenReturn("EMAIL")
-
-		activity.composeErrorEmail("url", TestException("error"))
-
-		assertThat(calledIntent?.action, `is`(Intent.ACTION_SENDTO))
-		assertThat(calledIntent?.data, `is`(Uri.parse("mailto:")))
-
-		val extras = calledIntent?.extras ?: Bundle()
-
-		assertThat(
-			extras.getStringArray(Intent.EXTRA_EMAIL)?.toList(),
-			hasItem("EMAIL")
-		)
-
-		assertThat(extras.getString(Intent.EXTRA_SUBJECT), `is`("SUBJECT"))
-
-		assertThat(
-			extras.getString(Intent.EXTRA_TEXT),
-			startsWith("url\n\nerror\n\ncom.darakeon.dfm.extensions.NavigationTest.composeErrorEmail")
-		)
-	}
-
-	@Test
-	fun composeErrorApi() {
-		`when`(activity.getString(R.string.error_call_api_email))
-			.thenReturn("API")
-
-		`when`(activity.getString(R.string.dfm_mail_address))
-			.thenReturn("EMAIL")
-
-		activity.composeErrorApi()
-
-		assertThat(calledIntent?.action, `is`(Intent.ACTION_SENDTO))
-		assertThat(calledIntent?.data, `is`(Uri.parse("mailto:")))
-
-		val extras = calledIntent?.extras ?: Bundle()
-
-		assertThat(
-			extras.getStringArray(Intent.EXTRA_EMAIL)?.toList(),
-			hasItem("EMAIL")
-		)
-
-		assertThat(extras.getString(Intent.EXTRA_SUBJECT), `is`("API"))
-		assertThat(extras.getString(Intent.EXTRA_TEXT), containsString("BaseActivity"))
 	}
 }
